Migrate runner/port-scanner.ts to typescript
