+++ .gitignore
@@ -6,3 +6,4 @@ typings/
 runner/browserrunner.js
 runner/paths.js
 runner/plugin.js
+runner/port-scanner.js
