+++ tsconfig.json
@@ -18,6 +18,7 @@
         "runner/browserrunner.ts",
         "runner/paths.ts",
         "runner/plugin.ts",
+        "runner/port-scanner.ts",
         "custom_typings/cleankill.d.ts",
         "custom_typings/promisify-node.d.ts",
         "custom_typings/wd.d.ts",
